Fixing issue with error when running XSLTs with DOCTYPE declaration.
When running an XSLT with a DOCTYPE declaration at the top. E.g.

```xsl
<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE stylesheet [
<!ENTITY cr "&#10;">
<!ENTITY tab "&#09;">
]>
<xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform"
...
```

Then I get the error: `DOCTYPE is disallowed when the feature "http://apache.org/xml/features/disallow-doctype-decl" set to true.`

There is a fix for this given at: http://stackoverflow.com/a/30754000

This is what has been applied here.